refactor(db/migrations): Did some cleanup to allow for a more cohesive… (#180)
* refactor(db/migrations: Did some cleanup to allow for a more cohesive organization of the db migration code.

This change was required to allow for a more organized codebase when it comes to the
organization of the database migration code. It was previously split between the
`internal/db.py` and the `config.py` files. I believe that was done due to causing circular dependencies
during the SQLAlchemy migration. This was caused by the Auth model being used to instantiate its table
by SQLAlchemy. The Auth class and AuthTable class were both in `models/auths.py` in which the AuthTable class
used the `get_db()` function to access the database engine. Since the DB engine isn't created before the completion
of the SQLAlchemy migration, this caused the circular dependency error.

To fix this, the Auth and AuthTable classes were split into seperate files to prevent this circular dependency.

* chore: fmt.

* fix: missed an import here.
