+++ backend/open_webui/models/channels.py
@@ -2,7 +2,8 @@ import time
 import uuid
 from typing import Optional
 
-from open_webui.internal.db import Base, get_db
+from open_webui.internal.db import get_db
+from open_webui.models.base import Base
 from open_webui.utils.access_control import has_access
 
 from pydantic import BaseModel, ConfigDict
