+++ backend/open_webui/models/feedbacks.py
@@ -3,7 +3,8 @@ import time
 import uuid
 from typing import Optional
 
-from open_webui.internal.db import Base, get_db
+from open_webui.internal.db import get_db
+from open_webui.models.base import Base
 from open_webui.models.users import User
 
 from open_webui.env import SRC_LOG_LEVELS
