+++ backend/open_webui/models/files.py
@@ -2,8 +2,9 @@ import logging
 import time
 from typing import Optional
 
-from open_webui.internal.db import Base, get_db
+from open_webui.internal.db import get_db
 from open_webui.env import SRC_LOG_LEVELS
+from open_webui.models.base import Base
 from pydantic import BaseModel, ConfigDict
 from sqlalchemy import BigInteger, Column, String, Text, JSON
 
@@ -197,6 +198,7 @@ class FilesTable:
                 file.data = {**(file.data if file.data else {}), **data}
                 db.commit()
                 return FileModel.model_validate(file)
+
             except Exception:
                 return None
