+++ backend/open_webui/models/folders.py
@@ -3,8 +3,9 @@ import time
 import uuid
 from typing import Optional
 
-from open_webui.internal.db import Base, get_db
+from open_webui.internal.db import get_db
 from open_webui.models.chats import Chats
+from open_webui.models.base import Base
 
 from open_webui.env import SRC_LOG_LEVELS
 from pydantic import BaseModel, ConfigDict
