+++ backend/open_webui/models/groups.py
@@ -3,9 +3,9 @@ import time
 from typing import Optional
 import uuid
 
-from open_webui.internal.db import Base, get_db
+from open_webui.internal.db import get_db
 from open_webui.env import SRC_LOG_LEVELS
-
+from open_webui.models.base import Base
 
 from pydantic import BaseModel, ConfigDict
 from sqlalchemy import BigInteger, Column, String, Text, JSON, func
