+++ backend/open_webui/models/knowledge.py
@@ -3,9 +3,9 @@ import time
 from typing import Optional
 import uuid
 
-from open_webui.internal.db import Base, get_db
+from open_webui.internal.db import get_db
 from open_webui.env import SRC_LOG_LEVELS
-
+from open_webui.models.base import Base
 from open_webui.models.files import FileMetadataResponse
 from open_webui.models.users import Users, UserResponse
