+++ backend/open_webui/models/message_metrics.py
@@ -3,7 +3,9 @@ from typing import Optional
 import uuid
 from pydantic import BaseModel
 from sqlalchemy import Column, Text, BigInteger, func
-from open_webui.internal.db import Base, get_db
+
+from open_webui.internal.db import get_db
+from open_webui.models.base import Base
 from logging import getLogger
 
 logger = getLogger(__name__)
