+++ backend/open_webui/models/messages.py
@@ -2,8 +2,8 @@ import time
 import uuid
 from typing import Optional
 
-from open_webui.internal.db import Base, get_db
-
+from open_webui.internal.db import get_db
+from open_webui.models.base import Base
 
 from pydantic import BaseModel, ConfigDict
 from sqlalchemy import BigInteger, Column, Text, JSON
