+++ backend/open_webui/models/models.py
@@ -2,9 +2,9 @@ import logging
 import time
 from typing import Optional
 
-from open_webui.internal.db import Base, JSONField, get_db
+from open_webui.internal.db import JSONField, get_db
 from open_webui.env import SRC_LOG_LEVELS
-
+from open_webui.models.base import Base
 from open_webui.models.users import Users, UserResponse
