+++ backend/open_webui/models/tags.py
@@ -1,10 +1,10 @@
 import logging
 from typing import Optional
 
-from open_webui.internal.db import Base, get_db
-
-
+from open_webui.internal.db import get_db
 from open_webui.env import SRC_LOG_LEVELS
+from open_webui.models.base import Base
+
 from pydantic import BaseModel, ConfigDict
 from sqlalchemy import Column, String, JSON, PrimaryKeyConstraint
